Update README.md
Fast and Furious is my favorite movie.  It is full of cars, action, and family moments.  The movie shows how a group of friends come together through racing and loyalty. It was first released in 2001 and became a big hit around the world.  The main character, Dominic Toretto, is a strong and loyal leader.  Brian O'Conner starts as a cop but becomes part of the family.  The movie also has amazing car stunts and races in different cities.  I like how every movie adds new characters and bigger action scenes.  Fast and Furious is not just about racing, it’s about trust, friendship, and family.